Rule removed from RuleSeq
Rule is now a class in the Language module rather than an inner class in RuleSeq. The Rule class has also been moved into a separate file.
